+++ src/main/java/org/sosly/witchcraft/Witchcraft.java
@@ -41,7 +41,7 @@ public class Witchcraft {
         MinecraftForge.EVENT_BUS.register(this);
 
         // Register our mod's ForgeConfigSpec so that Forge can create and load the config file for us
-        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, Config.SPEC);
+        ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, ServerConfig.SPEC);
 
         if (FMLEnvironment.dist.isClient()) {
             modbus.register(ScreenRegistry.class);