+++ src/main/java/org/sosly/witchcraft/enchantments/staves/DedicationEnchantment.java
@@ -9,7 +9,7 @@ import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
-import org.sosly.witchcraft.Config;
+import org.sosly.witchcraft.ServerConfig;
 
 public class DedicationEnchantment extends MAEnchantmentBase {
     private static final int chargeModifier = 40;
@@ -44,8 +44,8 @@ public class DedicationEnchantment extends MAEnchantmentBase {
 
         stack.getOrCreateTag().putInt("dedication", level);
 
-        float tierModifier = 1.0F + ((level - 1) * (float)Config.dedicationTierMultiplier);
-        float charges = Config.dedicationCharges * chargeModifier * tierModifier;
+        float tierModifier = 1.0F + ((level - 1) * (float)ServerConfig.dedicationTierMultiplier);
+        float charges = ServerConfig.dedicationCharges * chargeModifier * tierModifier;
         if (ItemUtils.getCharges(stack) == 0) {
             ItemUtils.writeCharges(stack, (int) charges);
         }